misc. improvements
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk/phpwiki@27 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ fpw/specialPages.php
@@ -1,109 +1,23 @@
 <?
-function error ( $error ) {
-	$page = new WikiPage ;
-	$page->special ( "Yikes! An error!" ) ;
-	$page->contents = "<h2>$error!</h2>Return to the [[:HomePage|HomePage]]!" ;
-	return $page->renderPage () ;
-	}
-
-function edit ( $title ) {
-	global $EditBox , $SaveButton , $PreviewButton , $MinorEdit ;
-	global $user , $CommentBox , $vpage , $EditTime ;
-	$npage = new WikiPage ;
-	$npage->title = $title ;
-	$npage->makeAll () ;
-	$ret = "" ;
-	if ( !$vpage->canEdit() ) return "<h3>You cannot edit this page!</h3>" ;
-
-	if ( $EditTime == "" ) $EditTime = date ( "YmdHis" ) ;
-
-
-	if ( isset ( $SaveButton ) ) {
-		unset ( $SaveButton ) ;
-		if ( $vpage->doesTopicExist() ) {
-			$lastTime = getMySQL ( "cur" , "cur_timestamp" , "cur_title=\"$vpage->secureTitle\"" ) ;
-			if ( tsc($EditTime) < tsc($lastTime) ) return "<h1>While you were typing, someone saved another version of this article!</h1>" ;
-			}
-		$text = $EditBox ;
-		$text = str_replace ( "\\'" , "'" , $text ) ;
-		$text = str_replace ( "\\\"" , "\"" , $text ) ;
-		if ( $user->isLoggedIn ) $text = str_replace ( "~~~" , "[[user:$user->name|$user->name]]" , $text ) ;
-		else $text = str_replace ( "~~~" , $user->getLink() , $text ) ;
-		$title = str_replace ( "\\'" , "'" , $title ) ;
-		$title = str_replace ( "\\\"" , "\"" , $title ) ;
-		$npage->title = $title ;
-		$npage->makeAll () ;
-		if ( $npage->doesTopicExist() ) $npage->backup() ;
-		else { $MinorEdit = 2 ; $npage->ensureExistence () ; }
-		if ( !$user->isLoggedIn ) $npage->setEntry ( $text , $CommentBox , 0 , $user->getLink() , $MinorEdit*1 ) ;
-		else $npage->setEntry ( $text , $CommentBox , $user->id , $user->name , $MinorEdit*1 ) ;
-		global $wasSaved ;
-		$wasSaved = true ;
-		return "" ;
-	} else if ( isset ( $PreviewButton ) ) {
-		unset ( $PreviewButton ) ;
-		$text = $EditBox ;
-		$text = str_replace ( "\\'" , "'" , $text ) ;
-		$text = str_replace ( "\\\"" , "\"" , $text ) ;
-		$append = "<hr>\n<h2>Preview :</h2>\n".$npage->parseContents($text)."<hr><h3>Remember, this is only a preview and not yet saved!</h3>" ;
-	} else if ( $npage->doesTopicExist() ) {
-		$npage->load ( $npage->title ) ;
-		$text = $npage->contents ;
-	} else {
-		$text = "Describe the new page here." ;
-		}
-
-	if ( $MinorEdit ) $checked = "checked" ;
-	else $checked = "" ;
-	if ( $CommentBox == "" ) $CommentBox = "*" ;
-
-
-	# Just trying to set the initial keyboard focus to the edit window; doesn't work, though...
-	global $bodyOptions , $headerScript ;
-	$headerScript = "<script> <!-- function setfocus() { document.f.EditBox.focus(); } --> </script>" ;
-	$bodyOptions = " onLoad=setfocus()" ;
-
-	$ret .= "<form method=POST name=f>" ;
-	$ret .= "<textarea tabindex=1 name=EditBox rows=".$user->options["rows"]." cols=".$user->options["cols"]." STYLE=\"width:100%\" WRAP=virtual>$text</textarea><br>\n" ;
-	$ret .= "Summary:<input tabindex=2 type=text value=\"$CommentBox\" name=CommentBox size=50 maxlength=200> \n" ;
-	$ret .= "<input tabindex=3 type=checkbox name=MinorEdit $checked value=1>This is a minor edit \n" ;
-	$ret .= "<input tabindex=4 type=submit value=Save name=SaveButton> \n" ;
-	$ret .= "<input tabindex=5 type=submit value=Preview name=PreviewButton>\n" ;
-	$ret .= "<input type=hidden value=\"$EditTime\" name=EditTime>\n" ;
-	$ret .= "</form>" ;
-
-	return $ret.$append ;
-	}
-
-function doEdit ( $title ) {
-	global $THESCRIPT ;
-	global $vpage , $action , $wasSaved ;
-	$wasSaved = false ;
-	$vpage = new WikiPage ;
-	$vpage->isSpecialPage = true ;
-	$vpage->title = $title ;
-	$vpage->makeSecureTitle () ;
-	$action = "" ;
-	$ret = $vpage->getHeader() ;
-	$action = "edit" ;
-
-	$theMiddle = edit ( $title ) ;
-	if ( $wasSaved )
-		$theMiddle = "<h1>Your page was successfully saved!</h1><META HTTP-EQUIV=Refresh CONTENT=\"0; URL=$THESCRIPT?title=$vpage->secureTitle\">" ;
-	$ret .= $vpage->getMiddle ( $theMiddle ) ;
-
-	$action = "" ;
-	$ret .= $vpage->getFooter() ;
-	$action = "edit" ;
-	return $ret ;
-	}
-
-function view ( $title ) {
-	global $vpage ;
-	$vpage = new WikiPage ;
-	$vpage->load ( $title ) ;
-	return $vpage->renderPage () ;
-	}
+# This contains special functions that are not necessare for basic view/edit purposes
+# - userLogout
+# - userLogin
+# - editUserSettings
+# - WantedPages (the most wanted)
+# - shortPages (stub articles)
+# - lonelyPages (orphans)
+# - watchlist
+# - recentChanges
+# - randomPage
+# - statistics
+# - allPages
+# - search
+# - specialPages (the list)
+# - history
+# - upload
+# - statistics
+# - delete (a page; for sysops only!)
+# - askSQL (for sysops only!)
 
 function userLogout () {
 	global $user , $vpage ;
@@ -497,55 +411,7 @@ function listUsers () {
 		if ( in_array ( "is_sysop" , $user->rights ) ) $ret .= " ($s->user_rights)" ;
 		$ret .= "\n" ;
 		}
-	
 
-	return $ret ;
-
-#------------------------------------------------
-	global $THESCRIPT ;
-	global $user , $vpage , $startat ;
-	if ( !isset ( $startat ) ) $startat = 1 ;
-	$perpage = $user->options["resultsPerPage"] ;
-	if ( $perpage == 0 ) $perpage = 20 ;
-	
-	$vpage->special ( "User List" ) ;
-	$vpage->namespace = "" ;
-	$ret = "'''These are all wikipedia users (that have an account)!'''\n\n" ;
-	$connection = getDBconnection () ;
-	mysql_select_db ( "wikipedia" , $connection ) ;
-	$sql = "SELECT COUNT(*) AS number FROM cur WHERE cur_title LIKE \"User:%\"" ;
-	$result = mysql_query ( $sql , $connection ) ;
-	$s = mysql_fetch_object ( $result ) ;
-	$total = $s->number ;
-	$sql = "SELECT * FROM cur WHERE cur_title LIKE \"User:%\" ORDER BY cur_title" ;
-	$result = mysql_query ( $sql , $connection ) ;
-	$cnt = 1 ;
-	$color1 = $user->options["tabLine1"] ;
-	$color2 = $user->options["tabLine2"] ;
-	$color = $color1 ;
-	$ret .= "<table width=100%>\n" ;
-	while ( $s = mysql_fetch_object ( $result ) and $cnt < $startat+$perpage ) {
-		if ( $cnt >= $startat ) {
-			$u = $s->cur_title ;
-			$u = substr ( strstr ( $u , ":" ) , 1 ) ;
-			$u = $vpage->getNiceTitle ( $u ) ;
-			$ret .= "<tr><td$color align=right nowrap>$cnt</td>" ;
-			$ret .= "<td$color width=100% valign=top>[[$s->cur_title|$u]]</td></tr>\n";
-			if ( $color == $color1 ) $color = $color2 ;
-			else $color = $color1 ;
-			}
-		$cnt++ ;
-		}
-	$ret .= "</table>\n" ;
-	
-	$ret .= "<nowiki>" ;
-	$before = $startat - $perpage ; $fin = $before + $perpage - 1 ;
-	if ( $startat > 1 ) $ret .= "<a href=\"$THESCRIPT?title=special:UserList&startat=$before\">$before-$fin&lt;&lt;</a> &nbsp;" ;
-	$after = $startat + $perpage ; $fin = $after+$perpage - 1 ; if ( $fin > $total ) $fin = $total ;
-	if ( $after-1 < $total ) $ret .= "<a href=\"$THESCRIPT?title=special:UserList&startat=$after\">&gt;&gt;$after-$fin</a>" ;
-	$ret .= "</nowiki>" ;
-	mysql_free_result ( $result ) ;
-	mysql_close ( $connection ) ;
 	return $ret ;
 	}
 
@@ -864,22 +730,14 @@ function upload () {
 	global $removeFile , $xtitle , $removeFile , $Upload , $Upload_name , $no_copyright ;
 	global $user , $vpage ;
 	$vpage->special ( "Upload Page" ) ;
-
-#	if ( $USERLOGGEDIN != "YES" ) return "You are not logged in! You have to be logged in to upload a file. <a href=\"$THESCRIPT?action=login\">Log in</a> or return to the <a href=\"$THESCRIPT?no\">HomePage</a>" ;
-
-#	$rights = ",".getUserSetting ( $USERNAME , "user_rights" )."," ;
-#	if ( strstr ( $rights , ",is_editor," ) or strstr ( $rights , ",is_sysop" ) ) $isEditor = true ;
-#	else $isEditor = false ;
-#	if ( strstr ( $rights , ",is_sysop," ) or strstr ( $rights , ",is_sysop" ) ) $isSysop = true ;
-#	else $isSysop = false ;
-
+	$isSysop = in_array ( "is_sysop" , $user->rights ) ;
 	$xtitle = "File upload page";
 	$ret = "<nowiki>" ;
 
 	$message = "" ;
 
 	if (isset($removeFile)) {
-#		if ( !$isSysop and !$isEditor ) return "You are neither an editor nor a sysop. Return to the <a href=\"$THESCRIPT?action=upload\">Upload page</a>" ;
+		if ( !$isSysop ) return "You are neither an editor nor a sysop. Return to the <a href=\"$THESCRIPT?action=upload\">Upload page</a>" ;
 		if (is_file("upload/$removeFile") ) unlink ("./upload/$removeFile");
 		$message = "File <b>$removeFile</b> deleted!" ;
 		unset ( $removeFile ) ;
@@ -949,16 +807,16 @@ function upload () {
 			$i = 0;
 			$ret .= "<table border=1 width=\"100%\">\n";
 			$ret .= "<tr><th>File</th><th>Size (byte)</th>";
-#			if ( $isSysop or $isEditor )
-			$ret .= "<th>File removal</th>";
+			if ( $isSysop )
+				$ret .= "<th>File removal</th>";
 			$ret .= "</tr>\n" ;
 			while ($entry = readdir($mydir)) {
 				if ($entry != '.' && $entry != '..') {
 					$ret .= "<tr><td align=center>" ;
 					$ret .= "<a href=upload/$entry>$entry</a></td>";
 					$ret .= "<td align=center>".filesize("upload/$entry")." bytes</td>";
-#					if ( $isSysop or $isEditor )
-					$ret .= "<td align=center><a href=\"$THESCRIPT?title=special:upload&removeFile=$entry\">Click here to remove $entry.</a></td>" ;
+					if ( $isSysop )
+						$ret .= "<td align=center><a href=\"$THESCRIPT?title=special:upload&removeFile=$entry\">Click here to remove $entry.</a></td>" ;
 					$ret .= "</tr>" ;
 					$i++;
 				}
@@ -1137,7 +995,7 @@ function deletepage () {
 	return "<nowiki>$ret</nowiki>" ;
 	}
 
-# A little hack; disabled; to enable, allow function call in wikiPage->load()
+# A little hack for direct MySQL access; for sysops only!
 function askSQL () {
 	global $THESCRIPT ;
 	global $Save , $question ;
